--- .github/workflows/boot-routes.yml
@@ -1,34 +1,0 @@
-name: spring-reactive-boot-routes
-
-on:
-  push:
-    paths:
-      - "boot-routes/**"
-    branches: [ master ]
-  pull_request:
-    paths:
-      - "boot-routes/**"
-    types:
-      - opened
-      - synchronize
-      - reopened
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-      with:
-          fetch-depth: 0
-    - name: Set up JDK
-      uses: actions/setup-java@v2.3.1
-      with:
-        java-version: '17'
-        distribution: 'adopt'
-        cache: 'maven'
-    - name: Build with Maven
-      run: |
-        cd boot-routes
-        mvn -B package --file pom.xml
