--- .github/workflows/boot-tomcat.yml
@@ -1,34 +1,0 @@
-name: spring-reactive-boot-tomcat
-
-on:
-  push:
-    paths:
-      - "boot-tomcat/**"
-    branches: [ master ]
-  pull_request:
-    paths:
-      - "boot-tomcat/**"
-    types:
-      - opened
-      - synchronize
-      - reopened
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-      with:
-          fetch-depth: 0
-    - name: Set up JDK
-      uses: actions/setup-java@v2.3.1
-      with:
-        java-version: '17'
-        distribution: 'adopt'
-        cache: 'maven'
-    - name: Build with Maven
-      run: |
-        cd boot-tomcat
-        mvn -B package --file pom.xml
