--- boot-jetty/src/main/resources/application.yml
@@ -1,13 +1,0 @@
-server:
-  port: 8080
-spring:
-  data:
-    mongodb:
-#      host: localhost
-#      port: 27017
-#      uri: mongodb://user:secret@mongo1.example.com:12345,mongo2.example.com:23456/test
-      uri: mongodb://localhost:27017/blog
-      grid-fs-database: images
-logging:
-  level:
-    org.springframework.web: DEBUG
