--- boot-routes/src/main/java/com/example/demo/DemoApplication.java
@@ -1,193 +1,0 @@
-package com.example.demo;
-
-import java.net.URI;
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.config.EnableMongoAuditing;
-import org.springframework.data.mongodb.core.mapping.Document;
-import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authorization.AuthorizationDecision;
-import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
-import org.springframework.security.config.web.server.ServerHttpSecurity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.MapReactiveUserDetailsService;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.server.SecurityWebFilterChain;
-import org.springframework.security.web.server.authorization.AuthorizationContext;
-import org.springframework.stereotype.Component;
-import static org.springframework.web.reactive.function.server.RequestPredicates.DELETE;
-import static org.springframework.web.reactive.function.server.RequestPredicates.GET;
-import static org.springframework.web.reactive.function.server.RequestPredicates.POST;
-import static org.springframework.web.reactive.function.server.RequestPredicates.PUT;
-import org.springframework.web.reactive.function.server.RouterFunction;
-import static org.springframework.web.reactive.function.server.RouterFunctions.route;
-import org.springframework.web.reactive.function.server.ServerRequest;
-import org.springframework.web.reactive.function.server.ServerResponse;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
-@SpringBootApplication
-@EnableMongoAuditing
-public class DemoApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DemoApplication.class, args);
-    }
-
-    @Bean
-    public RouterFunction<ServerResponse> routes(PostHandler postController) {
-        return route(GET("/posts"), postController::all)
-            .andRoute(POST("/posts"), postController::create)
-            .andRoute(GET("/posts/{id}"), postController::get)
-            .andRoute(PUT("/posts/{id}"), postController::update)
-            .andRoute(DELETE("/posts/{id}"), postController::delete);
-    }
-
-}
-
-@EnableWebFluxSecurity
-class SecurityConfig {
-
-    @Bean
-    SecurityWebFilterChain springWebFilterChain(ServerHttpSecurity http) throws Exception {
-        return http
-            .authorizeExchange()
-            .pathMatchers(HttpMethod.GET, "/posts/**").permitAll()
-            .pathMatchers(HttpMethod.DELETE, "/posts/**").hasRole("ADMIN")
-            //.pathMatchers("/users/{user}/**").access(this::currentUserMatchesPath)
-            .anyExchange().authenticated()
-            .and()
-            .build();
-    }
-
-    private Mono<AuthorizationDecision> currentUserMatchesPath(Mono<Authentication> authentication, AuthorizationContext context) {
-        return authentication
-            .map(a -> context.getVariables().get("user").equals(a.getName()))
-            .map(granted -> new AuthorizationDecision(granted));
-    }
-
-	@Bean
-	public MapReactiveUserDetailsService userDetailsRepository() {
-		UserDetails user = User.withDefaultPasswordEncoder().username("user").password("password").roles("USER").build();
-		UserDetails admin = User.withDefaultPasswordEncoder().username("admin").password("password").roles("USER","ADMIN").build();
-		return new MapReactiveUserDetailsService(user, admin);
-	}
-
-}
-
-@Component
-@Slf4j
-class DataInitializer implements CommandLineRunner {
-
-    private final PostRepository posts;
-
-    public DataInitializer(PostRepository posts) {
-        this.posts = posts;
-    }
-
-    @Override
-    public void run(String[] args) {
-        log.info("start data initialization  ...");
-        this.posts
-            .deleteAll()
-            .thenMany(
-                Flux
-                    .just("Post one", "Post two")
-                    .flatMap(
-                        title -> this.posts.save(Post.builder().title(title).content("content of " + title).build())
-                    )
-            )
-            .log()
-            .subscribe(
-                null,
-                null,
-                () -> log.info("done initialization...")
-            );
-
-    }
-
-}
-
-@Component
-class PostHandler {
-
-    private final PostRepository posts;
-
-    public PostHandler(PostRepository posts) {
-        this.posts = posts;
-    }
-
-    public Mono<ServerResponse> all(ServerRequest req) {
-        return ServerResponse.ok().body(this.posts.findAll(), Post.class);
-    }
-
-    public Mono<ServerResponse> create(ServerRequest req) {
-        return req.bodyToMono(Post.class)
-            .flatMap(post -> this.posts.save(post))
-            .flatMap(p -> ServerResponse.created(URI.create("/posts/" + p.getId())).build());
-    }
-
-    public Mono<ServerResponse> get(ServerRequest req) {
-        return this.posts.findById(req.pathVariable("id"))
-            .flatMap(post -> ServerResponse.ok().body(Mono.just(post), Post.class))
-            .switchIfEmpty(ServerResponse.notFound().build());
-    }
-
-    public Mono<ServerResponse> update(ServerRequest req) {
-
-        return Mono
-            .zip(
-                (data) -> {
-                    Post p = (Post) data[0];
-                    Post p2 = (Post) data[1];
-                    p.setTitle(p2.getTitle());
-                    p.setContent(p2.getContent());
-                    return p;
-                },
-                this.posts.findById(req.pathVariable("id")),
-                req.bodyToMono(Post.class)
-            )
-            .cast(Post.class)
-            .flatMap(post -> this.posts.save(post))
-            .flatMap(post -> ServerResponse.noContent().build());
-
-    }
-
-    public Mono<ServerResponse> delete(ServerRequest req) {
-        return ServerResponse.noContent().build(this.posts.deleteById(req.pathVariable("id")));
-    }
-
-}
-
-interface PostRepository extends ReactiveMongoRepository<Post, String> {
-}
-
-@Document
-@Data
-@ToString
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-class Post {
-
-    @Id
-    private String id;
-    private String title;
-    private String content;
-
-    @CreatedDate
-    private LocalDateTime createdDate;
-}
