+++ boot-start-routes/pom.xml
@@ -19,7 +19,7 @@
   </parent>
 
   <properties>
-    <java.version>11</java.version>
+    <java.version>17</java.version>
   </properties>
 
   <dependencies>
