+++ boot-start/src/main/java/com/example/demo/DemoApplication.java
@@ -1,15 +1,28 @@
 package com.example.demo;
 
+
+import java.net.URI;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 @SpringBootApplication
 public class DemoApplication {
@@ -20,24 +33,68 @@ public class DemoApplication {
 }
 
 @RestController
-@RequestMapping
-class MessageController {
+@RequestMapping(value = "/posts")
+class PostController {
 
-    @GetMapping
-    Flux<Message> allMessages(){
-        return Flux.just(
-            Message.builder().body("hello Spring 5").build(),
-            Message.builder().body("hello Spring Boot 2").build()
-        );
+    private final PostRepository posts;
+
+    public PostController(PostRepository posts) {
+        this.posts = posts;
     }
-   
+
+    @GetMapping(value = "")
+    public Flux<Post> all() {
+        return this.posts.findAll();
+    }
+
+    @GetMapping(value = "/{id}")
+    public Mono<Post> get(@PathVariable(value = "id") UUID id) {
+        return this.posts.findById(id);
+    }
+
+    @PostMapping(value = "")
+    public Mono<ResponseEntity<?>> create(@RequestBody Post post) {
+        return this.posts.save(post).map(p -> ResponseEntity.created(URI.create("/posts/" + p.getId())).build());
+    }
+
+}
+
+@Component
+@Slf4j
+class PostRepository {
+
+    private static final List<Post> DATA = new ArrayList<>();
+
+    static {
+        DATA.add(Post.builder().id(UUID.randomUUID()).title("post one").content("content of post one").build());
+        DATA.add(Post.builder().id(UUID.randomUUID()).title("post two").content("content of post two").build());
+    }
+
+    Flux<Post> findAll() {
+        return Flux.fromIterable(DATA);
+    }
+
+
+    Mono<Post> findById(UUID id) {
+        return findAll().filter(p -> p.getId().equals(id)).last();
+    }
+
+    Mono<Post> save(Post post) {
+        Post saved = Post.builder().id(UUID.randomUUID()).title(post.getTitle()).content(post.getContent()).build();
+        log.debug("saved post: {}", saved);
+        DATA.add(saved);
+        return Mono.just(saved);
+    }
+
 }
 
 @Data
+@ToString
 @Builder
-@AllArgsConstructor
 @NoArgsConstructor
-class Message {
-
-    String body;
+@AllArgsConstructor
+class Post {
+    private UUID id;
+    private String title;
+    private String content;
 }
