+++ boot-start/src/test/java/com/example/demo/DemoApplicationTests.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
-@WebFluxTest(controllers = MessageController.class)
+@WebFluxTest(controllers = PostController.class)
 public class DemoApplicationTests {
 
     @Autowired
