--- boot-tomcat/src/main/java/com/example/demo/DemoApplication.java
@@ -1,207 +1,0 @@
-package com.example.demo;
-
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.config.EnableMongoAuditing;
-import org.springframework.data.mongodb.core.mapping.Document;
-import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authorization.AuthorizationDecision;
-import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
-import org.springframework.security.config.web.server.ServerHttpSecurity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.MapReactiveUserDetailsService;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.server.SecurityWebFilterChain;
-import org.springframework.security.web.server.authorization.AuthorizationContext;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
-@SpringBootApplication
-@EnableMongoAuditing
-public class DemoApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DemoApplication.class, args);
-    }
-
-}
-
-
-/*
-@EnableWebFluxSecurity
-class SecurityConfiguration {
-
-    @Bean
-    UserDetailsRepository userDetailsRepository() {
-        return new MapReactiveUserDetailsService(user("rob").build(), user("josh").roles("USER","ADMIN").build());
-    }
-
-    private User.UserBuilder user(String username) {
-        return User.withDefaultPasswordEncoder().username(username).password("password").roles("USER");
-    }
-
-    @Bean
-    SecurityWebFilterChain springSecurity(ServerHttpSecurity http) {
-        return http
-                .authorizeExchange()
-                    .pathMatchers("/users/me").authenticated()
-                    .pathMatchers("/users/{username}").access((auth,context) ->
-                        auth
-                                .map( a-> a.getName().equals(context.getVariables().get("username")))
-                                .map(AuthorizationDecision::new)
-                    )
-                    .anyExchange().hasRole("ADMIN")
-                    .and()
-                .build();
-    }
-}*/
-
-@EnableWebFluxSecurity
-class SecurityConfig {
-
-	@Bean
-	SecurityWebFilterChain springWebFilterChain(ServerHttpSecurity http) throws Exception {
-		return http
-			.authorizeExchange()
-				.pathMatchers(HttpMethod.GET, "/posts/**").permitAll()
-                .pathMatchers(HttpMethod.DELETE, "/posts/**").hasRole("ADMIN")
-				//.pathMatchers("/users/{user}/**").access(this::currentUserMatchesPath)
-				.anyExchange().authenticated()
-				.and()
-			.build();
-	}
-
-	private Mono<AuthorizationDecision> currentUserMatchesPath(Mono<Authentication> authentication, AuthorizationContext context) {
-		return authentication
-			.map( a -> context.getVariables().get("user").equals(a.getName()))
-			.map( granted -> new AuthorizationDecision(granted));
-	}
-
-	@Bean
-	public MapReactiveUserDetailsService userDetailsRepository() {
-		UserDetails user = User.withDefaultPasswordEncoder().username("user").password("password").roles("USER").build();
-		UserDetails admin = User.withDefaultPasswordEncoder().username("admin").password("password").roles("USER","ADMIN").build();
-		return new MapReactiveUserDetailsService(user, admin);
-	}
-
-}
-
-
-@Component
-@Slf4j
-class DataInitializer implements CommandLineRunner {
-
-    private final PostRepository posts;
-
-    public DataInitializer(PostRepository posts) {
-        this.posts = posts;
-    }
-
-    @Override
-    public void run(String[] args) {
-        log.info("start data initialization  ...");
-        this.posts
-            .deleteAll()
-            .thenMany(
-                Flux
-                    .just("Post one", "Post two")
-                    .flatMap(
-                        title -> this.posts.save(Post.builder().title(title).content("content of " + title).build())
-                    )
-            )
-            .log()
-            .subscribe(
-                null,
-                null,
-                () -> log.info("done initialization...")
-            );
-
-    }
-
-}
-
-@RestController()
-@RequestMapping(value = "/posts")
-class PostController {
-
-    private final PostRepository posts;
-
-    public PostController(PostRepository posts) {
-        this.posts = posts;
-    }
-
-    @GetMapping("")
-    public Flux<Post> all() {
-        return this.posts.findAll();
-    }
-
-    @PostMapping("")
-    public Mono<Post> create(@RequestBody Post post) {
-        return this.posts.save(post);
-    }
-
-    @GetMapping("/{id}")
-    public Mono<Post> get(@PathVariable("id") String id) {
-        return this.posts.findById(id);
-    }
-
-    @PutMapping("/{id}")
-    public Mono<Post> update(@PathVariable("id") String id, @RequestBody Post post) {
-        return this.posts.findById(id)
-            .map(p -> {
-                p.setTitle(post.getTitle());
-                p.setContent(post.getContent());
-
-                return p;
-            })
-            .flatMap(p -> this.posts.save(p));
-    }
-
-    @DeleteMapping("/{id}")
-    public Mono<Void> delete(@PathVariable("id") String id) {
-        return this.posts.deleteById(id);
-    }
-
-}
-
-interface PostRepository extends ReactiveMongoRepository<Post, String> {
-}
-
-@Document
-@Data
-@ToString
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-class Post {
-
-    @Id
-    private String id;
-    private String title;
-    private String content;
-
-    @CreatedDate
-    private LocalDateTime createdDate;
-}
