--- boot-tomcat/src/test/java/com/example/demo/DemoApplicationTests.java
@@ -1,16 +1,0 @@
-package com.example.demo;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class DemoApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
